+++ src/Components/Loader.js
@@ -1,0 +1,6 @@
+import React from "react";
+const Loader = () => {
+  return <div className="loader">Loading...</div>;
+};
+
+export default Loader;
