+++ src/Pages/YoMorgenstern.js
@@ -67,6 +67,9 @@ function YoMorgen() {
     YoMorStyleDos.height = "1700px";
   }
 
+
+  
+
   return (
     <>
       <div style={{ backgroundColor: "black" }}>
