+++ src/Styles/App.css
@@ -1200,3 +1200,155 @@ left: 180px; */
   z-index: 990;
   bottom: 60px;
 }
+
+/* MOBILE */
+
+.mob-home-container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  height: 90vh;
+}
+.mob-img-cont {
+  display: flex;
+  justify-content: center;
+  margin-top: -82vh;
+}
+
+@media (max-width: 720px) {
+  .navbar-container {
+    display: flex;
+    flex-direction: column-reverse;
+    height: auto;
+  }
+  .navbar-title {
+    font-size: 33px;
+  }
+  .navbar-right-section {
+    display: flex;
+    justify-content: center;
+    align-items: flex-start;
+    flex-direction: column;
+    width: 100%;
+    margin-top: 100px;
+    margin-left: 40px;
+  }
+  .navbar-left-section {
+    width: 100%;
+    display: flex;
+    justify-content: space-around;
+    align-items: flex-start;
+    flex-direction: column-reverse;
+  }
+  .contact-me {
+    margin-left: 40px;
+    margin-top: 20px;
+  }
+  .contact {
+    margin-left: 0px;
+    color: black;
+    display: flex;
+    width: 100%;
+    justify-content: space-evenly;
+    margin-top: 30px;
+  }
+  .contact-icons-text {
+    display: none;
+  }
+  .contact-icons {
+    width: 20px;
+  }
+
+  .email-input {
+    /* height: 94%; */
+    /* width: 83.1%; */
+    border-radius: 0px;
+    /* height: 100% !important; */
+  }
+
+  .submit-button {
+    height: 112% !important;
+  }
+  .nav-bar {
+    margin-left: 0px;
+    top: 0px;
+    right: 0px;
+    width: 50vw;
+    display: flex;
+    justify-content: space-between;
+  }
+  .toggle-button {
+    position: absolute;
+    margin-left: 40px;
+    margin-top: 30px;
+  }
+  .plus-icon {
+    margin-top: 25px;
+    margin-right: 30px;
+  }
+}
+
+@media (max-width: 400px) {
+  .email-container {
+    width: 250px;
+    height: 45px;
+    position: relative;
+  }
+  .submit-button {
+    position: absolute;
+    right: -13px;
+    top: 2px;
+    width: 32%;
+  }
+}
+
+/* @media (max-height: 690px) {
+  .navbar-right-section {
+    margin-top: 0px;
+  }
+} */
+
+@media (max-height: 660px) {
+  .navbar-right-section {
+    margin-top: 90px;
+  }
+  .navbar-left-section {
+    /* margin-top: -0px; */
+  }
+
+  .navbar-title {
+    font-size: 29px;
+  }
+  .contact-me {
+    margin-top: 20px;
+  }
+  .contact {
+    margin-top: 15px;
+  }
+}
+
+@media (max-height: 610px) {
+  .navbar-right-section {
+    margin-top: 60px;
+  }
+  .navbar-left-section {
+    margin-top: -15px;
+  }
+
+  .navbar-title {
+    font-size: 29px;
+  }
+  .contact-me {
+    margin-top: 10px;
+  }
+  .contact {
+    margin-top: -5px;
+  }
+}
+
+@media (max-height: 530px) {
+  .contact-me {
+    display: none;
+  }
+}
